components/esp-matter: introduce the global command callback for the users
Added set/get APIs for the global user command callback. This gets
called twice before SDK processes the command and after SDK processed
the command.

Users can define the actions they need to perform. e.g. in case of
OnOff::On or OnOff::Off specification mandates moving the brightness to
min value and then transition to On or Off state. This has a flickering
effect and a lot more updates on the attributes, this may jam-up the
driver/cloud updates. If user has more knowledge of invoked command they
can device the stuff thay may need update and which ones to ignore.